chore: 식별값 Integer -> Long 으로 변경

+++ BE-MyCarMaster/src/main/java/softeer/be_my_car_master/api/color_interior/controller/InteriorColorController.java
@@ -26,8 +26,8 @@ public class InteriorColorController {
 	public Response<GetInteriorColorsResponse> getInterior(
 		@RequestBody @Valid GetInteriorColorsRequest getInteriorColorsRequest) {
 
-		Integer trimId = getInteriorColorsRequest.getTrimId();
-		Integer exteriorColorId = getInteriorColorsRequest.getExteriorColorId();
+		Long trimId = getInteriorColorsRequest.getTrimId();
+		Long exteriorColorId = getInteriorColorsRequest.getExteriorColorId();
 		GetInteriorColorsResponse getInteriorColorsResponse = getInteriorColorsUseCase.execute(trimId, exteriorColorId);
 		return Response.createSuccessResponse(getInteriorColorsResponse);
 	}
